+++ tests/unit/aggregation/test_imtl_g.py
@@ -1,4 +1,7 @@
 import pytest
+import torch
+from torch.testing import assert_close
+from unit.conftest import DEVICE
 
 from torchjd.aggregation import IMTLG
 
@@ -8,6 +11,21 @@ from ._property_testers import ExpectedShapeProperty, PermutationInvariancePrope
 @pytest.mark.parametrize("aggregator", [IMTLG()])
 class TestIMTLG(ExpectedShapeProperty, PermutationInvarianceProperty):
     pass
+
+
+def test_imtlg_zero():
+    """
+    Tests that IMTLG correctly returns the 0 vector in the special case where input matrix only
+    consists of zeros.
+    """
+
+    A = IMTLG()
+    J = torch.zeros(2, 3, device=DEVICE)
+
+    aggregation = A(J)
+    expected = torch.zeros(3, device=DEVICE)
+
+    assert_close(aggregation, expected)
 
 
 def test_representations():