+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). This changelog does not include internal
 changes that do not affect the user.
 
+## [Unreleased]
+
+### Fixed
+
+- **BREAKING** Prevented IMTLG from dividing by zero during its weight rescaling step. If the input
+  matrix consists only of zeros, it will now return a vector of zeros instead of a vector of `nan`.
+
 ## [0.2.0] - 2024-09-05
 
 ### Added